fix(core): fix critical circular dependencies
