+++ src/extensions/markdown/Image/index.ts
@@ -28,7 +28,6 @@ export const Image: ExtensionAuto<ImageOptions | undefined> = (builder, opts) =>
 declare global {
     namespace WysiwygEditor {
         interface Actions {
-            // @ts-expect-error
             [addImageAction]: Action<AddImageAttrs>;
         }
     }
