+++ src/extensions/yfm/ImgSize/index.ts
@@ -44,6 +44,7 @@ export const ImgSize: ExtensionAuto<ImgSizeOptions> = (builder, opts) => {
 declare global {
     namespace WysiwygEditor {
         interface Actions {
+            // @ts-expect-error
             [addImageAction]: Action<AddImageAttrs>;
         }
     }
