+++ src/toolbar/index.ts
@@ -5,4 +5,3 @@ export * from './ToolbarButton';
 export * from './ToolbarGroup';
 export * from './ToolbarListButton';
 export * from './FlexToolbar';
-export * from '../bundle/config';
